refactor: Rename cli options for extractor and template
Renamed the extractor selection from the cli to '--input' since it
decides the various input formats that are used to gather annotations
from.
Renamed the template selection from the cli to '--output' since it
control the output format that annotations are displayed/written in.

This also somewhat more closely mirrors pandoc cli options, which are
generally a good guide to follow.